python/vlm-service/requirements.txt
openvino>=2024.4.0
openvino-genai
openvino-tokenizers
transformers>=4.37.0
huggingface_hub>=0.20.0
Pillow>=9.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
thefuzz[speedup]>=0.22.0
numpy>=1.24.0
# Google Gemini engine for handwriting OCR
google-genai>=1.0.0
